Reorganize Admin -> Config controllers and templates
