+++ java/code/src/com/suse/manager/webui/Router.java
@@ -29,7 +29,8 @@ import static spark.Spark.notFound;
 import static spark.Spark.post;
 
 import com.suse.manager.webui.controllers.ActivationKeysController;
-import com.suse.manager.webui.controllers.AdminConfigController;
+import com.suse.manager.webui.controllers.admin.AdminApiController;
+import com.suse.manager.webui.controllers.admin.AdminViewsController;
 import com.suse.manager.webui.controllers.CVEAuditController;
 import com.suse.manager.webui.controllers.DownloadController;
 import com.suse.manager.webui.controllers.FormulaCatalogController;
@@ -98,6 +99,10 @@ public class Router implements SparkApplication {
         // Content Management Routes
         ContentManagementViewsController.initRoutes(jade);
         ContentManagementApiController.initRoutes();
+
+        // Admin Router
+        AdminViewsController.initRoutes(jade);
+        AdminApiController.initRoutes();
 
         // Minions
         get("/manager/systems/keys",
@@ -325,13 +330,6 @@ public class Router implements SparkApplication {
                 withProductAdmin(ProductsController::synchronizeSubscriptions));
         post("/manager/admin/setup/sync/repositories",
                 withProductAdmin(ProductsController::synchronizeRepositories));
-
-        // Admin - monitoring
-        get("/manager/admin/config/monitoring",
-                withUserPreferences(withCsrfToken(withOrgAdmin(AdminConfigController::showMonitoring))), jade);
-        get("/manager/api/admin/config/monitoring", withOrgAdmin(AdminConfigController::status));
-        post("/manager/api/admin/config/monitoring",
-                withOrgAdmin(AdminConfigController::changeMonitoringStatus));
     }
 
     private void  initNotFoundRoutes(JadeTemplateEngine jade) {
