+++ java/code/src/com/suse/manager/webui/controllers/admin/templates/monitoring.jade
